Rename helper -> diagnostic thx git
