+++ src/shared/diagnostic.cpp
